refactor: remove V2 suffix (from filenames)
