+++ packages/keyring-internal-snap-client/src/v2/KeyringInternalSnapClient.ts
