+++ packages/keyring-internal-snap-client/src/v2/index.ts
@@ -1,1 +1,1 @@
-export * from './KeyringInternalSnapClientV2';
+export * from './KeyringInternalSnapClient';
