+++ packages/keyring-snap-client/src/v2/KeyringClient.ts
