+++ packages/keyring-snap-client/src/v2/index.ts
@@ -1,1 +1,1 @@
-export * from './KeyringClientV2';
+export * from './KeyringClient';
